Integration/UI Testing for Android Studio Files of FillCart Use-Case
tests that add a banana menuItem to the cart database, checks the 2 existing components within activity_cart to make sure they are displayed, verifies that the item has been added by checking its name and price. Through these test it can be seen that the implementation of FillCart successfully is integrated with the other implemented use-cases.