Added links to video
These links are main on the HelpActivity, however in many other activities there are single TextViews that have now been linked up to also link to the video. The behaviour of these links and on click listeners is as follows; when the user clicks one the app tries to open the youtube android app and provide a link to said app, if this fails the app will open the phones browser onto the youtube page with the video. Each link except for ones reffering to create user have timestamps in their links to skip to the most important part of the video respective to where the button is or what the button's topic is supposed to encompass.